@@ -13,7 +13,7 @@
         }
     ],
      "require": {
-        "php": ">=5.6.0",
+        "php": ">=7.0",
         "silex/silex": "~2.0",
         "guzzle/guzzle": "*",
         "symfony/process": "~3",
