@@ -74,10 +74,9 @@ class RequestCollectionFacadeTest extends AbstractTestCase
     {
         $this->mockClient($path, $this->createResponseWithInvalidStatusCode(), $httpMethod);
 
-        $this->setExpectedException(
-            'UnexpectedValueException',
-            'Expected status code 200 from "' . $path . '", got 404'
-        );
+        $this->expectException('UnexpectedValueException');
+
+        $this->expectExceptionMessage('Expected status code 200 from "' . $path . '", got 404');
         call_user_func_array([$this->facade, $method], $args);
     }
 
@@ -86,10 +85,9 @@ class RequestCollectionFacadeTest extends AbstractTestCase
     {
         $this->mockClient($path, $this->createResponseWithEmptyContentType(), $httpMethod);
 
-        $this->setExpectedException(
-            'UnexpectedValueException',
-            'Expected content type "text/plain" from "' . $path . '", got ""'
-        );
+        $this->expectException('UnexpectedValueException');
+
+        $this->expectExceptionMessage('Expected content type "text/plain" from "' . $path . '", got ""');
         call_user_func_array([$this->facade, $method], $args);
     }
 
@@ -98,10 +96,9 @@ class RequestCollectionFacadeTest extends AbstractTestCase
     {
         $this->mockClient($path, $this->createResponseWithInvalidContentType(), $httpMethod);
 
-        $this->setExpectedException(
-            'UnexpectedValueException',
-            'Expected content type "text/plain" from "' . $path . '", got "text/html"'
-        );
+        $this->expectException('UnexpectedValueException');
+
+        $this->expectExceptionMessage('Expected content type "text/plain" from "' . $path . '", got "text/html"');
         call_user_func_array([$this->facade, $method], $args);
     }
 
@@ -110,10 +107,9 @@ class RequestCollectionFacadeTest extends AbstractTestCase
     {
         $this->mockClient($path, $this->createResponseThatCannotBeDeserialized(), $httpMethod);
 
-        $this->setExpectedException(
-            'UnexpectedValueException',
-            'Cannot deserialize response from "' . $path . '": "invalid response"'
-        );
+        $this->expectException('UnexpectedValueException');
+
+        $this->expectExceptionMessage('Cannot deserialize response from "' . $path . '": "invalid response"');
         call_user_func_array([$this->facade, $method], $args);
     }
 
